Ref: move config to resources folder
