+++ src/Resources/Config/footer.tsx
