+++ src/Resources/MockData/services.ts
@@ -3,66 +3,81 @@ import column from '../../Assets/images/column.png';
 import rehabilitation from '../../Assets/images/rehabilitation.png';
 import hands from '../../Assets/images/hands.png';
 import legs from '../../Assets/images/legs.png';
+
 import { ServicesCard } from '../../Components/AppCarousel/types';
 
-export const fullServicesItems: ServicesCard[] = [
+export const allClinicServices: ServicesCard[] = [
   {
-    id: 0,
+    id: 1,
     title: 'Cabeza',
+    shortDescription: 'Atención especializada para el cuidado y tratamiento de lesiones o afecciones en la cabeza.',
     description: `La fisioterapia para afecciones relacionadas con la cabeza es una rama especializada que aborda problemas como dolores cervicales, cefaleas tensionales, trastornos de la articulación temporomandibular (ATM), vértigo, mareos y secuelas de traumatismos craneoencefálicos. Los fisioterapeutas utilizan técnicas manuales, ejercicios terapéuticos y estiramientos para aliviar la tensión muscular, mejorar la movilidad cervical y reducir el dolor. En casos de vértigo o mareos, se emplean maniobras específicas, como la de Epley, para tratar problemas del sistema vestibular. Para trastornos de la ATM, se trabaja en la relajación de los músculos mandibulares y en la corrección de la alineación articular. Además, la fisioterapia puede incluir educación postural y ergonómica para prevenir recaídas y promover hábitos saludables. En pacientes con secuelas de traumatismos craneoencefálicos, el enfoque se centra en recuperar la funcionalidad, el equilibrio y la coordinación. La terapia es personalizada según las necesidades del paciente, combinando técnicas como la terapia manual, la punción seca o el uso de tecnología avanzada. La fisioterapia no solo busca aliviar los síntomas, sino también mejorar la calidad de vida y prevenir futuras complicaciones. Este enfoque integral ayuda a los pacientes a recuperar su autonomía y bienestar general.`,
     alt: 'Cabeza',
     image: brain,
-  },
-  {
-    id: 1,
-    title: 'Columna',
-    description: `La fisioterapia ofrece numerosos beneficios para la columna vertebral, que juega un papel crucial en la salud general y la movilidad. Uno de los principales beneficios es el alivio del dolor, ya que los fisioterapeutas utilizan diversas técnicas como la terapia manual, estiramientos y ejercicios para aliviar las molestias causadas por condiciones como la ciática, hernias discales o tensiones musculares. A través de ejercicios específicos, la fisioterapia ayuda a fortalecer los músculos que rodean la columna vertebral, proporcionando un mejor soporte y reduciendo la presión sobre las vértebras. Este soporte mejorado puede prevenir lesiones espinales adicionales y promover una mejor postura. Las sesiones regulares de fisioterapia también mejoran la flexibilidad y la movilidad articular, lo que es vital para mantener una columna vertebral saludable, especialmente a medida que envejecemos. Los fisioterapeutas enseñan métodos efectivos de corrección postural para reducir el riesgo de una mala alineación, un factor común en el dolor de espalda. Además, la fisioterapia ayuda en el proceso de rehabilitación después de cirugías o lesiones en la columna, acelerando la recuperación. Promueve la salud espinal general al enfocarse en patrones de movimiento correctos y fortalecer los músculos responsables de mantener la alineación espinal. Al abordar tanto los síntomas como las causas subyacentes de los problemas espinales, la fisioterapia respalda la salud espinal a largo plazo, ayudando a los pacientes a llevar una vida activa y sin dolor. Además, los fisioterapeutas proporcionan educación sobre cambios en el estilo de vida, como ajustes ergonómicos en el trabajo o durante las actividades diarias, lo que puede prevenir problemas recurrentes. El enfoque integral de la fisioterapia no solo aborda el dolor, sino que también mejora la calidad de vida de quienes sufren de molestias o limitaciones espinales.`,
-    alt: 'Columna',
-    image: column,
+    isPrimary: true,
   },
   {
     id: 2,
-    title: 'Brazos',
-    description: `La fisioterapia también ofrece grandes beneficios para los brazos, que son esenciales para una gran variedad de movimientos diarios y actividades físicas. Uno de los principales beneficios es el alivio del dolor, especialmente en casos de lesiones como esguinces, tendinitis o síndrome del túnel carpiano. A través de ejercicios terapéuticos específicos y técnicas de estiramiento, la fisioterapia ayuda a reducir la inflamación y mejorar la circulación sanguínea en los músculos y tejidos afectados, acelerando así la recuperación. Además, la fisioterapia contribuye a restaurar la fuerza y la flexibilidad de los músculos del brazo, permitiendo una mejor funcionalidad y evitando la debilidad que puede resultar de una lesión prolongada o inmovilización. El tratamiento fisioterapéutico también mejora la coordinación y la estabilidad de las articulaciones, como los codos, hombros y muñecas, reduciendo el riesgo de futuras lesiones. En casos de cirugía o lesiones graves, la fisioterapia acelera el proceso de rehabilitación, ayudando a restaurar la movilidad y el rango de movimiento de los brazos de manera progresiva y segura. Además, los fisioterapeutas enseñan ejercicios de prevención y técnicas de ergonomía que pueden aplicarse tanto en el trabajo como en las actividades diarias, reduciendo la probabilidad de que los problemas en los brazos se repitan. Gracias a estos tratamientos, los pacientes no solo mejoran su salud y funcionalidad, sino que también incrementan su calidad de vida al recuperar la capacidad para realizar actividades cotidianas y deportivas sin dolor. La fisioterapia es, por lo tanto, una herramienta clave para mantener los brazos fuertes, flexibles y saludables.`,
-    alt: 'Brazos',
-    image: hands,
+    title: 'Columna',
+    shortDescription: 'Tratamientos para problemas de la columna vertebral, como dolores, hernias y fracturas.',
+    description: `La fisioterapia ofrece numerosos beneficios para la columna vertebral, que juega un papel crucial en la salud general y la movilidad. Uno de los principales beneficios es el alivio del dolor, ya que los fisioterapeutas utilizan diversas técnicas como la terapia manual, estiramientos y ejercicios para aliviar las molestias causadas por condiciones como la ciática, hernias discales o tensiones musculares. A través de ejercicios específicos, la fisioterapia ayuda a fortalecer los músculos que rodean la columna vertebral, proporcionando un mejor soporte y reduciendo la presión sobre las vértebras. Este soporte mejorado puede prevenir lesiones espinales adicionales y promover una mejor postura. Las sesiones regulares de fisioterapia también mejoran la flexibilidad y la movilidad articular, lo que es vital para mantener una columna vertebral saludable, especialmente a medida que envejecemos. Los fisioterapeutas enseñan métodos efectivos de corrección postural para reducir el riesgo de una mala alineación, un factor común en el dolor de espalda. Además, la fisioterapia ayuda en el proceso de rehabilitación después de cirugías o lesiones en la columna, acelerando la recuperación. Promueve la salud espinal general al enfocarse en patrones de movimiento correctos y fortalecer los músculos responsables de mantener la alineación espinal. Al abordar tanto los síntomas como las causas subyacentes de los problemas espinales, la fisioterapia respalda la salud espinal a largo plazo, ayudando a los pacientes a llevar una vida activa y sin dolor. Además, los fisioterapeutas proporcionan educación sobre cambios en el estilo de vida, como ajustes ergonómicos en el trabajo o durante las actividades diarias, lo que puede prevenir problemas recurrentes. El enfoque integral de la fisioterapia no solo aborda el dolor, sino que también mejora la calidad de vida de quienes sufren de molestias o limitaciones espinales.`,
+    alt: 'Columna',
+    image: column,
+    isPrimary: true,
   },
   {
     id: 3,
+    title: 'Brazos',
+    shortDescription: 'Recuperación y tratamientos para lesiones en los brazos, muñecas y codos.',
+    description: `La fisioterapia también ofrece grandes beneficios para los brazos, que son esenciales para una gran variedad de movimientos diarios y actividades físicas. Uno de los principales beneficios es el alivio del dolor, especialmente en casos de lesiones como esguinces, tendinitis o síndrome del túnel carpiano. A través de ejercicios terapéuticos específicos y técnicas de estiramiento, la fisioterapia ayuda a reducir la inflamación y mejorar la circulación sanguínea en los músculos y tejidos afectados, acelerando así la recuperación. Además, la fisioterapia contribuye a restaurar la fuerza y la flexibilidad de los músculos del brazo, permitiendo una mejor funcionalidad y evitando la debilidad que puede resultar de una lesión prolongada o inmovilización. El tratamiento fisioterapéutico también mejora la coordinación y la estabilidad de las articulaciones, como los codos, hombros y muñecas, reduciendo el riesgo de futuras lesiones. En casos de cirugía o lesiones graves, la fisioterapia acelera el proceso de rehabilitación, ayudando a restaurar la movilidad y el rango de movimiento de los brazos de manera progresiva y segura. Además, los fisioterapeutas enseñan ejercicios de prevención y técnicas de ergonomía que pueden aplicarse tanto en el trabajo como en las actividades diarias, reduciendo la probabilidad de que los problemas en los brazos se repitan. Gracias a estos tratamientos, los pacientes no solo mejoran su salud y funcionalidad, sino que también incrementan su calidad de vida al recuperar la capacidad para realizar actividades cotidianas y deportivas sin dolor. La fisioterapia es, por lo tanto, una herramienta clave para mantener los brazos fuertes, flexibles y saludables.`,
+    alt: 'Brazos',
+    image: hands,
+    isPrimary: true,
+  },
+  {
+    id: 4,
     title: 'Piernas',
+    shortDescription: 'Soluciones para lesiones en las piernas, rodillas, tobillos y pies.',
     description: `La fisioterapia ofrece numerosos beneficios para las piernas, esenciales para el movimiento y la movilidad general del cuerpo. Uno de los principales beneficios es el alivio del dolor, especialmente en condiciones como artritis, esguinces, lesiones en los ligamentos o problemas musculares. A través de técnicas de terapia manual, estiramientos y ejercicios, los fisioterapeutas ayudan a reducir la inflamación y mejorar la circulación sanguínea en los músculos y tejidos afectados. Esto acelera el proceso de recuperación y mejora la función de las piernas. La fisioterapia también trabaja en el fortalecimiento de los músculos de las piernas, lo que proporciona una base más estable y reduce el riesgo de futuras lesiones. Los fisioterapeutas enseñan ejercicios específicos para mejorar la flexibilidad, el rango de movimiento y la coordinación de las articulaciones, como las rodillas, caderas y tobillos, lo que facilita una mayor movilidad y un caminar más fluido. En casos de rehabilitación después de una cirugía o una fractura, la fisioterapia es fundamental para restaurar la función completa de las piernas, ayudando a que el paciente recupere gradualmente la fuerza y el equilibrio. Además, los fisioterapeutas enseñan técnicas para caminar correctamente y mejorar la postura, lo cual es esencial para prevenir sobrecargas o lesiones por esfuerzos repetitivos. También brindan educación sobre hábitos de vida saludable y ergonomía para actividades como correr, caminar o estar de pie durante períodos prolongados. Con la fisioterapia, las personas pueden recuperar o mantener la capacidad de moverse de forma eficiente y sin dolor, mejorando así su calidad de vida y autonomía.`,
     alt: 'Piernas',
     image: legs,
   },
   {
-    id: 4,
+    id: 5,
     title: 'Rehabilitación',
+    shortDescription: 'Programas de rehabilitación para recuperar la movilidad y fuerza en cualquier parte del cuerpo.',
     description: `La rehabilitación es un proceso fundamental para recuperar la funcionalidad y calidad de vida después de una lesión, cirugía o enfermedad. A través de la fisioterapia, se facilita la restauración de la movilidad y el fortalecimiento de los músculos y articulaciones afectadas, ayudando a que el cuerpo recupere su capacidad de realizar actividades diarias sin dolor ni limitaciones. Uno de los principales beneficios de la rehabilitación es la mejora de la fuerza y la flexibilidad, lo que permite a los pacientes volver a realizar movimientos básicos como caminar, levantar objetos o realizar ejercicios deportivos. Los fisioterapeutas diseñan programas de ejercicios personalizados según las necesidades y limitaciones de cada persona, asegurando una recuperación gradual y segura. Además, la rehabilitación ayuda a reducir la inflamación, aliviar el dolor y mejorar la circulación sanguínea en las áreas afectadas, acelerando el proceso de recuperación. También se trabaja en la mejora del equilibrio, la coordinación y la postura, lo cual es vital para prevenir futuras lesiones. En casos de cirugías ortopédicas, neurológicas o cardiovasculares, la rehabilitación juega un papel clave en la recuperación funcional, ayudando a los pacientes a adaptarse a nuevos patrones de movimiento y mejorar su capacidad de realizar tareas cotidianas. La rehabilitación no solo se enfoca en la parte física, sino también en el bienestar emocional y psicológico del paciente, ayudando a recuperar la confianza en sí mismo y en sus capacidades. Este enfoque integral es esencial para una recuperación exitosa y para volver a una vida activa y saludable. Con el apoyo adecuado, la rehabilitación permite a las personas superar sus limitaciones y lograr una mejor calidad de vida a largo plazo.`,
     alt: 'Rehabilitación',
     image: rehabilitation,
+    isPrimary: true,
   },
   {
-    id: 5,
+    id: 6,
     title: 'Niños',
+    shortDescription: 'Fisioterapia especializada para bebés, niños y adolescentes.',
     description: `La fisioterapia para niños, también conocida como fisioterapia pediátrica, es una rama especializada de la terapia enfocada en mejorar la función física, la movilidad y el desarrollo general en bebés, niños pequeños y adolescentes. Aborda una amplia gama de afecciones, como retrasos en el desarrollo, trastornos congénitos, problemas neurológicos como la parálisis cerebral, afecciones musculoesqueléticas y lesiones. A través de ejercicios, estiramientos y actividades personalizadas, los fisioterapeutas ayudan a los niños a desarrollar fuerza, coordinación, equilibrio y habilidades motoras. La intervención temprana es crucial, ya que puede mejorar significativamente la capacidad del niño para alcanzar hitos del desarrollo y participar en actividades cotidianas. Los terapeutas suelen utilizar técnicas basadas en el juego para involucrar a los niños, haciendo que las sesiones sean divertidas y efectivas. Además, trabajan estrechamente con las familias, brindando orientación sobre ejercicios y estrategias para apoyar el progreso en casa. La fisioterapia pediátrica no solo mejora las habilidades físicas, sino que también aumenta la confianza y la independencia, permitiendo que los niños interactúen plenamente con su entorno y sus compañeros. Ya sea para abordar anomalías en la marcha, problemas posturales o la recuperación después de una cirugía, esta terapia desempeña un papel vital para garantizar que los niños lleven vidas más saludables y activas.`,
     alt: 'Niños',
     image: brain,
   },
   {
-    id: 6,
+    id: 7,
     title: 'Fisioterapia',
+    shortDescription:
+      'Somos profesionales expertos en el área de la fisioterapia, dedicados y comprometidos a garantizar una mejoría general de tu estado de salud y calidad de vida.',
     description: `
-Somos profesionales expertos en el área de la fisioterapia, dedicados y comprometidos a garantizar una mejoría general de tu estado de salud y calidad de vida.
-Es por ello, que a pesar de que somos un centro de Fisioterapia, principalmente, trabajamos bajo un equipo multidisciplinar que involucra: La nutrición, preparación física, psicología, entre otros. Para poder garantizar un abordaje integral y ayudarte a solucionar cualquier tipo de problema.
-A pesar de ser Fisioterapéutas, somos concientes de que nuestra labor se basa en ir más allá del abordaje de los síntomas de una patología. Más bien, hacemos énfasis en crear conciencia con respecto a cada uno de los factores que han podido originar la patología o lesión, para no solo solucionarla desde raíz, sino también para prevenir su aparición a futuro.
-Ponte en contacto con nosotros y comienza a disfrutar de un estilo de vida más saludable.
-`,
+  Somos profesionales expertos en el área de la fisioterapia, dedicados y comprometidos a garantizar una mejoría general de tu estado de salud y calidad de vida.
+  Es por ello, que a pesar de que somos un centro de Fisioterapia, principalmente, trabajamos bajo un equipo multidisciplinar que involucra: La nutrición, preparación física, psicología, entre otros. Para poder garantizar un abordaje integral y ayudarte a solucionar cualquier tipo de problema.
+  A pesar de ser Fisioterapéutas, somos concientes de que nuestra labor se basa en ir más allá del abordaje de los síntomas de una patología. Más bien, hacemos énfasis en crear conciencia con respecto a cada uno de los factores que han podido originar la patología o lesión, para no solo solucionarla desde raíz, sino también para prevenir su aparición a futuro.
+  Ponte en contacto con nosotros y comienza a disfrutar de un estilo de vida más saludable.
+  `,
     alt: 'Fisioterapia',
     image: column,
   },
   {
-    id: 7,
+    id: 8,
     title: 'Multidisciplinaria',
+    shortDescription:
+      'Enfoque integral que involucra a profesionales de diversas disciplinas para abordar de manera completa las necesidades de los pacientes en su proceso de recuperación.',
     description: `La rehabilitación multidisciplinaria es un enfoque integral que involucra a profesionales de diversas disciplinas para abordar de manera completa las necesidades de los pacientes en su proceso de recuperación. Este enfoque reconoce que los problemas de salud, especialmente aquellos derivados de lesiones o enfermedades complejas, afectan no solo al cuerpo, sino también a aspectos emocionales, sociales y psicológicos. En un equipo multidisciplinario, intervienen fisioterapeutas, médicos, psicólogos, nutricionistas, terapeutas ocupacionales, y otros especialistas, quienes colaboran para ofrecer una atención holística. El objetivo es garantizar que todos los aspectos de la salud del paciente sean tratados de manera coordinada, optimizando así los resultados del tratamiento. Por ejemplo, los fisioterapeutas pueden enfocarse en mejorar la movilidad física, mientras que los psicólogos abordan el bienestar emocional y el manejo del dolor, y los nutricionistas trabajan en la mejora de la alimentación para promover una mejor recuperación. Además, este enfoque permite personalizar los planes de tratamiento de acuerdo con las necesidades individuales de cada paciente, lo que aumenta la efectividad de la rehabilitación. La rehabilitación multidisciplinaria también facilita una recuperación más rápida y completa, ya que se minimizan las complicaciones y se previenen recaídas al abordar tanto los síntomas físicos como los factores subyacentes. Al trabajar en equipo, los profesionales pueden compartir información y coordinar sus esfuerzos para proporcionar un plan de atención más eficaz y centrado en el paciente. En resumen, la rehabilitación multidisciplinaria no solo acelera la recuperación, sino que también mejora la calidad de vida de los pacientes al abordar su bienestar de manera integral.`,
     alt: 'Multidisciplinaria',
     image: hands,
